requirements.txt
jinja2==3.0.3
python-dotenv
mljar-mercury
Pillow<7
torch
torchvision